analyzer: Remove dynamic call from checker
I managed to remove 2 of ~5 calls to _elementType, as the argument
type was known to have a property, `.type`. I removed the null check
from the function, as it never needed to be used; the argument should
always be non-null. Two calls then require null checks, but this does
not change behavior, because null was never passed in there; `dynamic`
would have been returned then, and cast to FunctionType, which would
have failed.

Change-Id: I9b93a11e60e20a99ba8df4b5677e4c5796a49fcc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/191384
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>