Migration: clarify error messages when preview server doesn't respond.
Previously, if the preview server couldn't be reached because it had
exited, we would generate an error message with an obscure looking
stacktrace.  I believe this was prompting users to file bug reports
against the migration tool, when in fact all they needed to do was
restart the migration process.  Hopefully this better error message
should help avoid confusion.

I also took the opportunity to improve the error message that is shown
if the user does restart the migration tool but tries to keep
interacting with it using an old authToken.

Fixes #45202.

Bug: https://github.com/dart-lang/sdk/issues/45202
Change-Id: I83e0ca99d789739e614bfff2d535d427a2fdfbb2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/192732
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>